move common scale functions out of line to axis
